+++ changelog.md
@@ -1,7 +1,13 @@
 # Change Logs
 
-## [1.13.2] - 27/05/2026
+## [1.13.3] - 31/05/2026
+
+### Added
+- Parsing of number literals, string literals, and identifiers is now interned and stored in `sys.interns`.
+- Warning messages in the parser about the use of the `is` and `is not` operators on number and string literals.
+- _etc._
 
 ### Fixed
+- Improved lexer and highlighting on numbers.
 - Fixed some bugs.
 - _etc._